fix(build): Apply a workaround for building Comrak in Nix
Carnix can not be configured to not build a dependency's
binaries (even if relevant features are disabled), but Comrak's binary
can by default not be built in Carnix because it attempts to read a
Cargo-provided environment variable at compile time.

This works around the issue by "faking" the environment variable,
which is less than ideal but works.